move bundle specific logic to main.rs
